+++ previous-versions/commerce/src/sdk/sdk.ts
@@ -61,8 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.37.0";
-    genVersion = "2.131.1";
+    sdkVersion = "0.38.0";
+    genVersion = "2.139.1";
+    userAgent = "speakeasy-sdk/typescript 0.38.0 2.139.1 3.0.0 @codat/commerce";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);